Conflicts:
	game.h
	monmove.cpp
	player.cpp
	pldata.h

Also added martial arts/techniques json data + character creation

+++ game.cpp
@@ -27,7 +27,7 @@
 #include "mapdata.h"
 #include "catacharset.h"
 #include "translations.h"
-#include "item.h"
+#include "init.h"
 #include <map>
 #include <set>
 #include <algorithm>
@@ -36,7 +36,11 @@
 #include <sstream>
 #include <math.h>
 #include <vector>
-#ifndef _MSC_VER
+
+#ifdef _MSC_VER
+#include "wdirent.h"
+#include <direct.h>
+#else
 #include <unistd.h>
 #include <dirent.h>
 #endif
@@ -45,6 +49,9 @@
 #include "artifactdata.h"
 
 #if (defined _WIN32 || defined __WIN32__)
+#ifndef NOMINMAX
+#define NOMINMAX
+#endif
 #include <windows.h>
 #include <tchar.h>
 #endif
@@ -83,17 +90,14 @@ game::game() :
  if(!json_good())
   throw (std::string)"Failed to initialize a static variable";
  // Gee, it sure is init-y around here!
+    load_json_dir("data/raw"); // load it, load it all!
  init_npctalk();
- init_materials();
  init_artifacts();
  init_weather();
  init_overmap();
  init_fields();
  init_faction_data();
  init_morale();
- init_skills();
- init_professions();
- init_bionics();              // Set up bionics                   (SEE bionics.cpp)
  init_mtypes();               // Set up monster types             (SEE mtypedef.cpp)
  init_techniques();           // Set up techniques                (SEE martialarts.cpp)
  init_itypes();               // Set up item types                (SEE itypedef.cpp)
@@ -139,18 +143,6 @@ game::~game()
  delwin(w_location);
  delwin(w_status);
  delwin(w_status2);
-}
-
-void game::init_skills() throw (std::string)
-{
-    try
-    {
-    Skill::skills = Skill::loadSkills();
-    }
-    catch (std::string &error_message)
-    {
-        throw;
-    }
 }
 
 // Fixed window sizes
@@ -5025,27 +5017,37 @@ void game::explosion(int x, int y, int power, int shrapnel, bool has_fire)
 
 void game::flashbang(int x, int y, bool player_immune)
 {
- int dist = rl_dist(u.posx, u.posy, x, y), t;
- if (dist <= 8 && !player_immune) {
-  if (!u.has_bionic("bio_ears"))
-   u.add_disease("deaf", 40 - dist * 4);
-  if (m.sees(u.posx, u.posy, x, y, 8, t))
-   u.infect("blind", bp_eyes, (12 - dist) / 2, 10 - dist, this);
- }
- for (int i = 0; i < num_zombies(); i++) {
-  monster &z = _z[i];
-  dist = rl_dist(z.posx(), z.posy(), x, y);
-  if (dist <= 4)
-   z.add_effect(ME_STUNNED, 10 - dist);
-  if (dist <= 8) {
-   if (z.has_flag(MF_SEES) && m.sees(z.posx(), z.posy(), x, y, 8, t))
-    z.add_effect(ME_BLIND, 18 - dist);
-   if (z.has_flag(MF_HEARS))
-    z.add_effect(ME_DEAF, 60 - dist * 4);
-  }
- }
- sound(x, y, 12, _("a huge boom!"));
-// TODO: Blind/deafen NPC
+    g->draw_explosion(x, y, 8, c_white);
+    int dist = rl_dist(u.posx, u.posy, x, y), t;
+    if (dist <= 8 && !player_immune) {
+        if (!u.has_bionic("bio_ears")) {
+            u.add_disease("deaf", 40 - dist * 4);
+        }
+        if (m.sees(u.posx, u.posy, x, y, 8, t)) {
+            int flash_mod = 0;
+            if (u.has_bionic("bio_sunglasses")) {
+                flash_mod = 6;
+            }
+            u.infect("blind", bp_eyes, (12 - flash_mod - dist) / 2, 10 - dist, this);
+        }
+    }
+    for (int i = 0; i < num_zombies(); i++) {
+        monster &z = _z[i];
+        dist = rl_dist(z.posx(), z.posy(), x, y);
+        if (dist <= 4) {
+            z.add_effect(ME_STUNNED, 10 - dist);
+        }
+        if (dist <= 8) {
+            if (z.has_flag(MF_SEES) && m.sees(z.posx(), z.posy(), x, y, 8, t)) {
+                z.add_effect(ME_BLIND, 18 - dist);
+            }
+            if (z.has_flag(MF_HEARS)) {
+                z.add_effect(ME_DEAF, 60 - dist * 4);
+            }
+        }
+    }
+    sound(x, y, 12, _("a huge boom!"));
+    // TODO: Blind/deafen NPC
 }
 
 void game::shockwave(int x, int y, int radius, int force, int stun, int dam_mult, bool ignore_player)
@@ -5676,6 +5678,26 @@ void game::clear_zombies()
     z_at.clear();
 }
 
+/**
+ * Attempts to spawn a hallucination somewhere close to the player. Returns
+ * false if the hallucination couldn't be spawned for whatever reason, such as
+ * a monster already in the target square.
+ * @return Whether or not a hallucination was successfully spawned.
+ */
+bool game::spawn_hallucination()
+{
+  monster phantasm(mtypes[rng(1, num_monsters - 1)]);
+  phantasm.hallucination = true;
+  phantasm.spawn(u.posx + rng(-10, 10), u.posy + rng(-10, 10));
+
+  //Don't attempt to place phantasms inside of other monsters
+  if (mon_at(phantasm.posx(), phantasm.posy()) == -1) {
+    return add_zombie(phantasm);
+  } else {
+    return false;
+  }
+}
+
 int game::mon_at(const int x, const int y) const
 {
     std::map<point, int>::const_iterator i = z_at.find(point(x, y));
@@ -5739,8 +5761,9 @@ void game::kill_mon(int index, bool u_did_it)
     mdeath tmpdeath;
     tmpdeath.guilt(this, &z);
    }
-   if (z.type->species != species_hallu)
+   if (!z.is_hallucination()) {
     kills[z.type->id]++;	// Increment our kill counter
+   }
   }
   for (int i = 0; i < z.inv.size(); i++)
    m.add_item_or_charges(z.posx(), z.posy(), z.inv[i]);
@@ -5757,6 +5780,10 @@ void game::explode_mon(int index)
   return;
  }
  monster &z = _z[index];
+ if(z.is_hallucination()) {
+   //Can't gib hallucinations
+   return;
+ }
  if (!z.dead) {
   z.dead = true;
   kills[z.type->id]++;	// Increment our kill counter
@@ -6085,7 +6112,7 @@ void game::use_item(char chInput)
 {
  char ch;
  if (chInput == '.')
-  ch = inv_activatable(_("Use item:"));
+  ch = inv(_("Use item:"));
  else
   ch = chInput;
 
@@ -6422,9 +6449,7 @@ void game::control_vehicle()
     vehicle *veh = m.veh_at(u.posx, u.posy, veh_part);
 
     if (veh && veh->player_in_control(&u)) {
-        std::string message = veh->use_controls();
-        if (!message.empty())
-            add_msg(message.c_str());
+        veh->use_controls();
     } else if (veh && veh->part_with_feature(veh_part, "CONTROLS") >= 0
                    && u.in_vehicle) {
         u.controlling_vehicle = true;
@@ -6442,9 +6467,7 @@ void game::control_vehicle()
             add_msg(_("No controls there."));
             return;
         }
-        std::string message = veh->use_controls();
-        if (!message.empty())
-            add_msg(message.c_str());
+        veh->use_controls();
     }
 }
 
@@ -8485,15 +8508,16 @@ void game::plfire(bool burst)
   return;
  }
  if (u.weapon.has_flag("CHARGE") && !u.weapon.active) {
-  if (u.has_charges("UPS_on", 1) || u.has_charges("UPS_off", 1) ||
-      u.has_charges("adv_UPS_on", 1) || u.has_charges("adv_UPS_off", 1)) {
+  if (u.has_charges("UPS_on", 1)  ||
+      u.has_charges("adv_UPS_on", 1) ) {
    add_msg(_("Your %s starts charging."), u.weapon.tname().c_str());
    u.weapon.charges = 0;
+   u.weapon.poison = 0;
    u.weapon.curammo = dynamic_cast<it_ammo*>(itypes["charge_shot"]);
    u.weapon.active = true;
    return;
   } else {
-   add_msg(_("You need a charged UPS."));
+   add_msg(_("You need a powered UPS."));
    return;
   }
  }
@@ -8845,7 +8869,7 @@ void game::eat(char chInput)
   return;
  }
  if (chInput == '.')
-  ch = inv_type(_("Consume item:"), IC_COMESTIBLE );
+  ch = inv_type(_("Consume item:"), IC_COMESTIBLE);
  else
   ch = chInput;
 
@@ -9311,7 +9335,7 @@ void game::plmove(int dx, int dy)
      monster &z = zombie(mondex);
      if (z.friendly == 0) {
          int udam = u.hit_mon(this, &z);
-         if (z.hurt(udam)) {
+         if (z.hurt(udam) || z.is_hallucination()) {
              kill_mon(mondex, true);
          }
          draw_hit_mon(x,y,z,z.dead);
@@ -9717,7 +9741,8 @@ void game::plmove(int dx, int dy)
       add_msg(_("There are vehicle controls here.  %s to drive."),
               press_x(ACTION_CONTROL_VEHICLE).c_str() );
 
- } else if (!m.has_flag(swimmable, x, y) && u.has_active_bionic("bio_probability_travel")) { //probability travel through walls but not water
+  } else if (!m.has_flag(swimmable, x, y) && u.has_active_bionic("bio_probability_travel") && u.power_level >= 10) {
+  //probability travel through walls but not water
   int tunneldist = 0;
   // tile is impassable
   while((m.move_cost(x + tunneldist*(x - u.posx), y + tunneldist*(y - u.posy)) == 0 &&
@@ -10399,54 +10424,58 @@ void game::update_stair_monsters()
 
 void game::despawn_monsters(const bool stairs, const int shiftx, const int shifty)
 {
- for (unsigned int i = 0; i < num_zombies(); i++) {
-  monster &z = zombie(i);
-  // If either shift argument is non-zero, we're shifting.
-  if(shiftx != 0 || shifty != 0) {
-   z.shift(shiftx, shifty);
-   if (z.posx() >= 0 - SEEX             && z.posy() >= 0 - SEEX &&
-       z.posx() <= SEEX * (MAPSIZE + 1) && z.posy() <= SEEY * (MAPSIZE + 1))
-     // We're inbounds, so don't despawn after all.
-     continue;
-  }
+    for (unsigned int i = 0; i < num_zombies(); i++) {
+        monster &z = zombie(i);
+        // If either shift argument is non-zero, we're shifting.
+        if(shiftx != 0 || shifty != 0) {
+            z.shift(shiftx, shifty);
+            if( z.posx() >= 0 && z.posx() <= SEEX * MAPSIZE &&
+                z.posy() >= 0 && z.posy() <= SEEY * MAPSIZE ) {
+                // We're inbounds, so don't despawn after all.
+                continue;
+            }
+        }
 
-  if (stairs && z.will_reach(this, u.posx, u.posy)) {
-   int turns = z.turns_to_reach(this, u.posx, u.posy);
-   if (turns < 999)
-    coming_to_stairs.push_back( monster_and_count(z, 1 + turns) );
-  } else if ( (z.spawnmapx != -1) ||
-      ((stairs || shiftx != 0 || shifty != 0) && z.friendly != 0 ) ) {
-    // translate shifty relative coordinates to submapx, submapy, subtilex, subtiley
-    real_coords rc(levx, levy, z.posx(), z.posy()); // this is madness
-    z.spawnmapx = rc.sub.x;
-    z.spawnmapy = rc.sub.y;
-    z.spawnposx = rc.sub_pos.x;
-    z.spawnposy = rc.sub_pos.y;
+        if (stairs && z.will_reach(this, u.posx, u.posy)) {
+            int turns = z.turns_to_reach(this, u.posx, u.posy);
+            if (turns < 999) {
+                coming_to_stairs.push_back( monster_and_count(z, 1 + turns) );
+            }
+        } else if ( (z.spawnmapx != -1) ||
+                    ((stairs || shiftx != 0 || shifty != 0) && z.friendly != 0 ) ) {
+            // translate shifty relative coordinates to submapx, submapy, subtilex, subtiley
+            real_coords rc(levx, levy, z.posx(), z.posy()); // this is madness
+            z.spawnmapx = rc.sub.x;
+            z.spawnmapy = rc.sub.y;
+            z.spawnposx = rc.sub_pos.x;
+            z.spawnposy = rc.sub_pos.y;
 
-    tinymap tmp(&traps);
-    tmp.load(this, z.spawnmapx, z.spawnmapy, levz, false);
-    tmp.add_spawn(&z);
-    tmp.save(cur_om, turn, z.spawnmapx, z.spawnmapy, levz);
-  } else {
-   	// No spawn site, so absorb them back into a group.
-   int group = valid_group((mon_id)(z.type->id), levx + shiftx, levy + shifty, levz);
-   if (group != -1) {
-    cur_om->zg[group].population++;
-    if (cur_om->zg[group].population / (cur_om->zg[group].radius * cur_om->zg[group].radius) > 5 &&
-        !cur_om->zg[group].diffuse)
-     cur_om->zg[group].radius++;
-   }
-  }
-  // Shifting needs some cleanup for despawned monsters since they won't be cleared afterwards.
-  if(shiftx != 0 || shifty != 0) {
-    remove_zombie(i);
-    i--;
-  }
- }
+            tinymap tmp(&traps);
+            tmp.load(this, z.spawnmapx, z.spawnmapy, levz, false);
+            tmp.add_spawn(&z);
+            tmp.save(cur_om, turn, z.spawnmapx, z.spawnmapy, levz);
+        } else {
+            // No spawn site, so absorb them back into a group.
+            int group = valid_group((mon_id)(z.type->id), levx + shiftx, levy + shifty, levz);
+            if (group != -1) {
+                cur_om->zg[group].population++;
+                if (cur_om->zg[group].population /
+                    (cur_om->zg[group].radius * cur_om->zg[group].radius) > 5 &&
+                    !cur_om->zg[group].diffuse) {
+                    cur_om->zg[group].radius++;
+                }
+            }
+        }
+        // Shifting needs some cleanup for despawned monsters since they won't be cleared afterwards.
+        if(shiftx != 0 || shifty != 0) {
+            remove_zombie(i);
+            i--;
+        }
+    }
 
- // The order in which zombies are shifted may cause zombies to briefly exist on
- // the same square. This messes up the mon_at cache, so we need to rebuild it.
- rebuild_mon_at_cache();
+    // The order in which zombies are shifted may cause zombies to briefly exist on
+    // the same square. This messes up the mon_at cache, so we need to rebuild it.
+    rebuild_mon_at_cache();
 }
 
 void game::spawn_mon(int shiftx, int shifty)
